[FIX] sale_stock_picking_return_fees_invoicing: Rename to sale_stock_restocking_fee_invoicing and use better naming into the code
