--- sale_stock_picking_return_fees_invoicing/README.rst
@@ -1,102 +1,0 @@
-========================================
-Sale Stock Picking Return Fees Invoicing
-========================================
-
-.. !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-   !! This file is generated by oca-gen-addon-readme !!
-   !! changes will be overwritten.                   !!
-   !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-
-.. |badge1| image:: https://img.shields.io/badge/maturity-Beta-yellow.png
-    :target: https://odoo-community.org/page/development-status
-    :alt: Beta
-.. |badge2| image:: https://img.shields.io/badge/licence-AGPL--3-blue.png
-    :target: http://www.gnu.org/licenses/agpl-3.0-standalone.html
-    :alt: License: AGPL-3
-.. |badge3| image:: https://img.shields.io/badge/github-OCA%2Faccount--invoicing-lightgray.png?logo=github
-    :target: https://github.com/OCA/account-invoicing/tree/10.0/sale_stock_picking_return_fees_invoicing
-    :alt: OCA/account-invoicing
-.. |badge4| image:: https://img.shields.io/badge/weblate-Translate%20me-F47D42.png
-    :target: https://translation.odoo-community.org/projects/account-invoicing-10-0/account-invoicing-10-0-sale_stock_picking_return_fees_invoicing
-    :alt: Translate me on Weblate
-.. |badge5| image:: https://img.shields.io/badge/runbot-Try%20me-875A7B.png
-    :target: https://runbot.odoo-community.org/runbot/95/10.0
-    :alt: Try me on Runbot
-
-|badge1| |badge2| |badge3| |badge4| |badge5| 
-
-This module extends the functionality of the stock picking return wizard
-to allows you to charge return fees in case of return of goods by a
-customer.
-
-**Table of contents**
-
-.. contents::
-   :local:
-
-Configuration
-=============
-
-The product used to charge return fees on the sale order must be specified
-into the stock settings form.
-
-Usage
-=====
-
-Once installed, A check box is added on the stock return wizard to specify if
-fees must be charged. The same field is added on each line into the wizard.
-You are therefore able to choice to charge for all the goods returned or select
-which line should be charged.
-
-By default no fees are charged in case of customer return. I you want to
-charge fees you can:
-
-* Specify on the customer contact form that fees must be charged by default
-  in case of return of goods. This value is used as default value into the
-  stock return wizard;
-* Use the checkboxes into the stock return wizard.
-
-For each returned product, if fees must be charged, once the return picking
-is done, a new line is added into the origin sale order to charge fees. The
-quantity charged is always 1 whatever the number of items of the same product
-returned.
-
-Bug Tracker
-===========
-
-Bugs are tracked on `GitHub Issues <https://github.com/OCA/account-invoicing/issues>`_.
-In case of trouble, please check there if your issue has already been reported.
-If you spotted it first, help us smashing it by providing a detailed and welcomed
-`feedback <https://github.com/OCA/account-invoicing/issues/new?body=module:%20sale_stock_picking_return_fees_invoicing%0Aversion:%2010.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
-
-Do not contact contributors directly about support or help with technical issues.
-
-Credits
-=======
-
-Authors
-~~~~~~~
-
-* ACSONE SA/NV
-
-Contributors
-~~~~~~~~~~~~
-
-* Laurent Mignon <laurent.mignon@acsone.eu> (https://acsone.eu)
-
-Maintainers
-~~~~~~~~~~~
-
-This module is maintained by the OCA.
-
-.. image:: https://odoo-community.org/logo.png
-   :alt: Odoo Community Association
-   :target: https://odoo-community.org
-
-OCA, or the Odoo Community Association, is a nonprofit organization whose
-mission is to support the collaborative development of Odoo features and
-promote its widespread use.
-
-This module is part of the `OCA/account-invoicing <https://github.com/OCA/account-invoicing/tree/10.0/sale_stock_picking_return_fees_invoicing>`_ project on GitHub.
-
-You are welcome to contribute. To learn how please visit https://odoo-community.org/page/Contribute.
