--- sale_stock_picking_return_fees_invoicing/__manifest__.py
@@ -1,24 +1,0 @@
-# -*- coding: utf-8 -*-
-# Copyright 2020 ACSONE SA/NV
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
-
-{
-    "name": "Sale Stock Picking Return Fees Invoicing",
-    "summary": """
-        On demand apply return fees on customer return""",
-    "version": "10.0.1.0.0",
-    "license": "AGPL-3",
-    "author": "ACSONE SA/NV,Odoo Community Association (OCA)",
-    "website": "https://acsone.eu/",
-    "depends": ["sale", "sale_stock", "stock", "stock_account"],
-    "data": [
-        "data/product_product.xml",
-        "views/res_company.xml",
-        "views/res_partner.xml",
-        "views/stock_move.xml",
-        "views/stock_picking.xml",
-        "wizards/stock_config_settings.xml",
-        "wizards/stock_return_picking.xml",
-    ],
-    "demo": [],
-}
