--- sale_stock_picking_return_fees_invoicing/data/product_product.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo noupdate="1">
-
-    <record id="product_customer_return_fees" model="product.product">
-        <field name="name">Customer return fees</field>
-        <field name="list_price">100</field>
-        <field name="type">service</field>
-        <field name="sale_ok" eval="False"/>
-        <field name="invoice_policy">order</field>
-    </record>
-
-</odoo>
