--- sale_stock_picking_return_fees_invoicing/models/sale_order_line.py
@@ -1,12 +1,0 @@
-# -*- coding: utf-8 -*-
-# Copyright 2020 ACSONE SA/NV
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
-
-from odoo import fields, models
-
-
-class SaleOrderLine(models.Model):
-
-    _inherit = 'sale.order.line'
-
-    is_customer_return_fees = fields.Boolean()
