--- sale_stock_picking_return_fees_invoicing/readme/CONFIGURE.rst
@@ -1,2 +1,0 @@
-The product used to charge return fees on the sale order must be specified
-into the stock settings form.
