--- sale_stock_picking_return_fees_invoicing/readme/DESCRIPTION.rst
@@ -1,3 +1,0 @@
-This module extends the functionality of the stock picking return wizard
-to allows you to charge return fees in case of return of goods by a
-customer.
