--- sale_stock_picking_return_fees_invoicing/readme/USAGE.rst
@@ -1,17 +1,0 @@
-Once installed, A check box is added on the stock return wizard to specify if
-fees must be charged. The same field is added on each line into the wizard.
-You are therefore able to choice to charge for all the goods returned or select
-which line should be charged.
-
-By default no fees are charged in case of customer return. I you want to
-charge fees you can:
-
-* Specify on the customer contact form that fees must be charged by default
-  in case of return of goods. This value is used as default value into the
-  stock return wizard;
-* Use the checkboxes into the stock return wizard.
-
-For each returned product, if fees must be charged, once the return picking
-is done, a new line is added into the origin sale order to charge fees. The
-quantity charged is always 1 whatever the number of items of the same product
-returned.
