--- sale_stock_picking_return_fees_invoicing/views/res_company.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!-- Copyright 2020 ACSONE SA/NV
-     License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl). -->
-
-<odoo>
-
-    <record model="ir.ui.view" id="res_company_form_view">
-        <field name="name">res.company.form (in sale_stock_picking_return_fees_invoicing)</field>
-        <field name="model">res.company</field>
-        <field name="inherit_id" ref="base.view_company_form"/>
-        <field name="arch" type="xml">
-            <field name="sequence" position="after">
-                <field name="customer_return_fees_product_id"/>
-            </field>
-        </field>
-    </record>
-</odoo>
