--- sale_stock_picking_return_fees_invoicing/views/res_partner.xml
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!-- Copyright 2020 ACSONE SA/NV
-     License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl). -->
-
-<odoo>
-
-    <record model="ir.ui.view" id="res_partner_form_view">
-        <field name="name">res.partner.form (in sale_stock_picking_return_fees_invoicing)</field>
-        <field name="model">res.partner</field>
-        <field name="inherit_id" ref="base.view_partner_form"/>
-        <field name="arch" type="xml">
-            <xpath expr="//group[@name='sale']/field[@name='user_id']" position="after">
-                <field name="charge_customer_stock_return"/>
-            </xpath>
-        </field>
-    </record>
-
-
-</odoo>
