--- sale_stock_picking_return_fees_invoicing/views/stock_move.xml
@@ -1,34 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!-- Copyright 2020 ACSONE SA/NV
-     License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl). -->
-
-<odoo>
-
-    <record model="ir.ui.view" id="stock_move_form_view">
-        <field name="name">stock.move.form (in sale_stock_picking_return_fees_invoicing)</field>
-        <field name="model">stock.move</field>
-        <field name="inherit_id" ref="stock.view_move_form"/>
-        <field name="arch" type="xml">
-            <field name="date_expected" position="after">
-                <field name="is_customer_return" invisible="1"/>
-                <field name="charge_customer_stock_return"  attrs="{'invisible':[('is_customer_return','=',False)]}"/>
-            </field>
-        </field>
-    </record>
-
-
-    <record model="ir.ui.view" id="stock_move_picking_form_view">
-        <field name="name">stock.move.picking.form (in sale_stock_picking_return_fees_invoicing)</field>
-        <field name="model">stock.move</field>
-        <field name="inherit_id" ref="stock.view_move_picking_form"/>
-        <field name="arch" type="xml">
-            <field name="date_expected" position="after">
-                <field name="is_customer_return" invisible="1"/>
-                <field name="charge_customer_stock_return"  attrs="{'invisible':[('is_customer_return','=',False)]}"/>
-            </field>
-        </field>
-    </record>
-
-
-
-</odoo>
