--- sale_stock_picking_return_fees_invoicing/views/stock_picking.xml
@@ -1,29 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!-- Copyright 2020 ACSONE SA/NV
-     License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl). -->
-
-<odoo>
-
-    <record model="ir.ui.view" id="stock_picking_form_view">
-        <field name="name">stock.return.picking.form (in
-            sale_stock_picking_return_fees_invoicing)
-        </field>
-        <field name="model">stock.picking</field>
-        <field name="inherit_id" ref="stock.view_picking_form"/>
-        <field name="arch" type="xml">
-            <!-- In order to be able to conditionally display the 'charge_customer_stock_return'
-            into the product_return_moves tree of the stock.return.picking form we declare into
-            action context that the return wizard is used into the context of a customer return.
-            This is required since it's not possible in Odoo 10 to makes a column visible into
-            a tree according to value of the parent record. In Odoo 12, this hack should be
-            replace by using the column_invisible attribute -->
-            <button name="%(stock.act_stock_return_picking)d" position="before">
-                 <field name="is_customer_out" invisible="1"/>
-            </button>
-            <button name="%(stock.act_stock_return_picking)d" position="attributes">
-                <attribute name="context">{"is_customer_return": is_customer_out}</attribute>
-            </button>
-        </field>
-    </record>
-
-</odoo>
