--- sale_stock_picking_return_fees_invoicing/wizards/stock_config_settings.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!-- Copyright 2020 ACSONE SA/NV
-     License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl). -->
-
-<odoo>
-
-    <record model="ir.ui.view" id="stock_config_settings_form_view">
-        <field name="name">stock.config.settings.form (in sale_stock_picking_return_fees_invoicing)</field>
-        <field name="model">stock.config.settings</field>
-        <field name="inherit_id" ref="stock_account.view_stock_config_settings_inherit"/>
-        <field name="arch" type="xml">
-            <field name="group_stock_inventory_valuation" position="after">
-                <field name="customer_return_fees_product_id"/>
-            </field>
-        </field>
-    </record>
-
-</odoo>
