+++ sale_stock_restocking_fee_invoicing/models/sale_order_line.py
@@ -1,0 +1,12 @@
+# -*- coding: utf-8 -*-
+# Copyright 2020 ACSONE SA/NV
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+
+from odoo import fields, models
+
+
+class SaleOrderLine(models.Model):
+
+    _inherit = 'sale.order.line'
+
+    is_restocking_fee = fields.Boolean()
