+++ sale_stock_restocking_fee_invoicing/readme/CONFIGURE.rst
@@ -1,0 +1,2 @@
+The product used to charge restocking fee on the sale order must be specified
+into the stock settings form.
