+++ sale_stock_restocking_fee_invoicing/readme/DESCRIPTION.rst
@@ -1,0 +1,3 @@
+This module extends the functionality of the stock picking return wizard
+to allows you to charge restocking fee in case of return of goods by a
+customer.
