+++ sale_stock_restocking_fee_invoicing/readme/ROADMAP.rst
@@ -1,0 +1,5 @@
+* Add support for different restocking fee policy. In the current
+  implementation a fixed price by returned products line is charged. It would
+  be nice to be able to select an other restocking fee policy. For example
+  the restocking fee could be based on a percentage of the value of goods
+  returned.
