+++ sale_stock_restocking_fee_invoicing/readme/USAGE.rst
@@ -1,0 +1,17 @@
+Once installed, A check box is added on the stock return wizard to specify if
+restocking fee must be charged. The same field is added on each line into
+the wizard. You are therefore able to choice to charge for all the goods
+returned or select which line should be charged.
+
+By default no restocking fee are charged in case of customer return. If you
+want to charge restocking fee you can:
+
+* Specify on the customer contact form that restocking fee must be charged by
+  default in case of return of goods. This value is used as default value into
+  the stock return wizard;
+* Use the checkboxes into the stock return wizard.
+
+For each returned product, if restocking fee must be charged, once the return
+picking is done, a new line is added into the origin sale order to charge fees.
+The quantity charged is always 1 whatever the number of items of the same
+product returned.
