--- setup/sale_stock_picking_return_fees_invoicing/odoo/addons/sale_stock_picking_return_fees_invoicing
@@ -1,1 +1,0 @@
-../../../../sale_stock_picking_return_fees_invoicing
