+++ setup/sale_stock_restocking_fee_invoicing/odoo/addons/__init__.py
