+++ setup/sale_stock_restocking_fee_invoicing/odoo/addons/sale_stock_restocking_fee_invoicing
@@ -1,0 +1,1 @@
+../../../../sale_stock_restocking_fee_invoicing
